+++ .task/index.js
@@ -1,9 +1,13 @@
 'use strict';
 
-var bootstrap = require('./tasks/modules/bootstrap');
 var bower = require('./tasks/bower');
 var build = require('./tasks/build');
 var help = require('./tasks/help');
+var modul = {
+  bootstrap:  require('./tasks/modules/bootstrap'),
+  _string:    require('./tasks/modules/underscore-string'),
+  rivets:     require('./tasks/modules/rivets')
+};
 var scaffold = {
   model:          require('./tasks/scaffolds/model'),
   collection:     require('./tasks/scaffolds/collection'),
@@ -17,24 +21,44 @@ module.exports = {
   add: {
     bootstrap: {
       description:  'Add Bootstrap',
-      task:         function(){ bootstrap.add() },
+      task:         function(){ modul.bootstrap.add() },
 
       js: {
-        description:  'Add Bootstrap JavaScript\n',
-        task:         function(){ bootstrap.js.add() }
+        description:  'Add Bootstrap JavaScript',
+        task:         function(){ modul.bootstrap.js.add() }
       }
+    },
+    underscore: {
+      string: {
+        description:  'Add Understore String',
+        task:         function(){ modul._string.add() }
+      }
+    },
+    rivets: {
+      description:  'Add Rivets for data binding between views and models\n',
+      task:         function(){ modul.rivets.add() }
     }
   },
 
   rem: {
     bootstrap: {
       description:  'Remove Bootstrap and any of the above Bootstrap-related',
-      task:         function(){ bootstrap.remove() },
+      task:         function(){ modul.bootstrap.remove() },
 
       js: {
-        description:  'Remove Bootstrap JavaScript\n',
-        task:         function(){ bootstrap.js.remove() }
+        description:  'Remove Bootstrap JavaScript',
+        task:         function(){ modul.bootstrap.js.remove() }
       }
+    },
+    underscore: {
+      string: {
+        description:  'Remove Understore String',
+        task:         function(){ modul._string.remove() }
+      }
+    },
+    rivets: {
+      description:  'Remove Rivets\n',
+      task:         function(){ modul.rivets.remove() }
     }
   },
 
@@ -69,28 +93,28 @@ module.exports = {
   del: {
     model: {
       description:  'Delete a Chaplin Model',
-      task:         function(){ scaffold.model.generate() }
+      task:         function(){ scaffold.model.destroy() }
     },
     view: {
       description:  'Delete a Chaplin View',
-      task:         function(){ scaffold.view.generate() },
+      task:         function(){ scaffold.view.destroy() },
 
       collection: {
         description:  'Delete a Chaplin CollectionView + Item View',
-        task:         function(){ scaffold.collectionView.generate() }
+        task:         function(){ scaffold.collectionView.destroy() }
       }
     },
     collection: {
       description:  'Delete a Chaplin Collection + Model',
-      task:         function(){ scaffold.collection.generate() }
+      task:         function(){ scaffold.collection.destroy() }
     },
     controller: {
       description:  'Delete a Chaplin Controller',
-      task:         function(){ scaffold.controller.generate() }
+      task:         function(){ scaffold.controller.destroy() }
     },
     test: {
       description:  'Delete a Mocha test file\n',
-      task:         function(){ scaffold.test.generate() }
+      task:         function(){ scaffold.test.destroy() }
     }
   },