+++ .task/tasks/modules/bootstrap.js
@@ -1,6 +1,5 @@
 'use strict';
 
-var Q = require('q');
 var modul = require('../../lib/modul');
 var bootstrap = module.exports = Object.create(modul);